feat: SendViaLinkDataService
