+++ p2p_wallet/Services/SendViaLinkStorage/SendViaLinkStorage.swift
@@ -4,84 +4,6 @@ import SolanaSwift
 import SwiftyUserDefaults
 import Combine
 
-// MARK: - TransactionInfo
-
-struct SendViaLinkTransactionInfo: Codable, Identifiable {
-    let amount: Double
-    let amountInFiat: Double
-    let token: Token
-    let seed: String
-    let timestamp: Date
-    
-    var id: String {
-        seed
-    }
-    
-    var creationDayInString: String {
-        // if today
-        if Calendar.current.isDateInToday(timestamp) {
-            return L10n.today
-        }
-        
-        // if another day
-        else {
-            let dateFormatter = DateFormatter()
-            dateFormatter.dateFormat = "MMMM d, yyyy"
-            let someDateString = dateFormatter.string(from: timestamp)
-            return someDateString
-        }
-    }
-    
-    var creationTimeInString: String {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "HH:mm"
-        let someDateString = dateFormatter.string(from: timestamp)
-        return someDateString
-    }
-}
-
-#if DEBUG
-extension Array where Element == SendViaLinkTransactionInfo {
-    static var mocked: Self {
-        [
-            .init(
-                amount: 10,
-                amountInFiat: 20,
-                token: .nativeSolana,
-                seed: .generateSendViaLinkSeed(),
-                timestamp: Date() // Today
-            ),
-            .init(
-                amount: 1,
-                amountInFiat: 0.99,
-                token: .usdc,
-                seed: .generateSendViaLinkSeed(),
-                timestamp: Date()
-                    .addingTimeInterval(-60*60*24*1) // This time yesterday
-            ),
-            .init(
-                amount: 1,
-                amountInFiat: 1.01,
-                token: .usdt,
-                seed: .generateSendViaLinkSeed(),
-                timestamp: Date()
-                    .addingTimeInterval(-60*60*24*2) // 2 days ago
-            ),
-            .init(
-                amount: 100,
-                amountInFiat: 1,
-                token: .srm,
-                seed: .generateSendViaLinkSeed(),
-                timestamp: Date()
-                    .addingTimeInterval(-60*60*24*1) // 3 days ago
-            )
-        ]
-    }
-}
-#endif
-
-// MARK: - Protocols
-
 /// Storage that handle SendViaLinkTransactions
 protocol SendViaLinkStorage {
